Refactoring to make filters and Flow app object oriented
