+++ flow/blocks/filters/__init__.py
@@ -2,10 +2,15 @@ import abc
 import base64
 import cStringIO
 from PIL import Image
+from ..block import Block
+# from blocks.block import Block
 
 
-class Filter(object):
+class Filter(Block):
     __metaclass__ = abc.ABCMeta
+
+    def __init__(self, block_spec):
+        super(Filter, self).__init__(block_spec)
 
     @abc.abstractmethod
     def compute(self, inputs, params):
