+++ flow/blocks/filters/brightness.py
@@ -4,6 +4,8 @@ from rhizo.extensions.camera import encode_image
 
 
 class Brightness(Filter):
+    def __init__(self, block_spec):
+        super(Brightness, self).__init__(block_spec)
 
     def compute(self, inputs, params):
         image_string = inputs[0]
